src/functions/requirements.txt
azure-functions==1.20.0
azure-storage-blob==12.19.0
requests==2.32.3
certifi>=2025.7.14 
